chore: preparation for refactoring
